+++ packages/insomnia-inso/package.json
@@ -62,7 +62,6 @@
     "consola": "^2.15.0",
     "cosmiconfig": "^6.0.0",
     "enquirer": "^2.3.6",
-    "insomnia-plugin-cookie-jar": "^3.19.0-beta.4",
     "insomnia-plugin-core-themes": "^3.19.0-beta.4",
     "insomnia-plugin-default-headers": "^3.19.0-beta.4",
     "insomnia-plugin-prompt": "^3.19.0-beta.4",
