--- plugins/insomnia-plugin-cookie-jar/README.md
@@ -1,3 +1,0 @@
-# Insomnia Cookie Jar Template Tag
-
-This is a core Insomnia plugin.
