--- plugins/insomnia-plugin-cookie-jar/package.json
@@ -1,33 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-plugin-cookie-jar",
-  "version": "3.19.0-beta.4",
-  "author": "Kong <office@konghq.com>",
-  "contributors": [
-    {
-      "name": "Preston Alvarado",
-      "email": "pjalva1@hotmail.com"
-    }
-  ],
-  "description": "Insomnia cookie jar template tag",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "plugins/insomnia-plugin-cookie-jar"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "index.js",
-  "insomnia": {
-    "name": "cookieJar",
-    "unlisted": true
-  },
-  "scripts": {
-    "test": "jest --silent"
-  },
-  "dependencies": {
-    "tough-cookie": "^4.1.3"
-  }
-}
